Refactored the source code. Removed Sandbox project.
