+++ Starcounter.Linq.sln
@@ -1,11 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.27004.2002
+VisualStudioVersion = 15.0.27004.2009
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Starcounter.Linq", "src\Starcounter.Linq.csproj", "{DF08E065-AAF1-42DB-A92E-D288154F5663}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sandbox", "test\Sandbox\Sandbox.csproj", "{471F5BC5-FB1F-4D76-9D06-DE76D943BE44}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Demo", "test\Demo\Demo.csproj", "{3F282C79-FC3D-482C-8287-7471594A060E}"
 EndProject
@@ -37,14 +35,6 @@ Global
 		{DF08E065-AAF1-42DB-A92E-D288154F5663}.Release|Any CPU.Build.0 = Release|Any CPU
 		{DF08E065-AAF1-42DB-A92E-D288154F5663}.Release|x64.ActiveCfg = Release|Any CPU
 		{DF08E065-AAF1-42DB-A92E-D288154F5663}.Release|x64.Build.0 = Release|Any CPU
-		{471F5BC5-FB1F-4D76-9D06-DE76D943BE44}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{471F5BC5-FB1F-4D76-9D06-DE76D943BE44}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{471F5BC5-FB1F-4D76-9D06-DE76D943BE44}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{471F5BC5-FB1F-4D76-9D06-DE76D943BE44}.Debug|x64.Build.0 = Debug|Any CPU
-		{471F5BC5-FB1F-4D76-9D06-DE76D943BE44}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{471F5BC5-FB1F-4D76-9D06-DE76D943BE44}.Release|Any CPU.Build.0 = Release|Any CPU
-		{471F5BC5-FB1F-4D76-9D06-DE76D943BE44}.Release|x64.ActiveCfg = Release|Any CPU
-		{471F5BC5-FB1F-4D76-9D06-DE76D943BE44}.Release|x64.Build.0 = Release|Any CPU
 		{3F282C79-FC3D-482C-8287-7471594A060E}.Debug|Any CPU.ActiveCfg = Debug|x64
 		{3F282C79-FC3D-482C-8287-7471594A060E}.Debug|Any CPU.Build.0 = Debug|x64
 		{3F282C79-FC3D-482C-8287-7471594A060E}.Debug|x64.ActiveCfg = Debug|x64
@@ -74,7 +64,6 @@ Global
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{471F5BC5-FB1F-4D76-9D06-DE76D943BE44} = {62519F4C-D4CE-4C59-A594-378E06AEFDEA}
 		{3F282C79-FC3D-482C-8287-7471594A060E} = {62519F4C-D4CE-4C59-A594-378E06AEFDEA}
 		{F39F0151-A278-47BB-8D5F-4A627590E487} = {62519F4C-D4CE-4C59-A594-378E06AEFDEA}
 		{1F9A5F6E-D0DE-4738-BF11-F1827989680E} = {62519F4C-D4CE-4C59-A594-378E06AEFDEA}
