+++ src/Visitors/OrderByVisitor.cs
@@ -3,7 +3,7 @@ using Starcounter.Linq.Helpers;
 
 namespace Starcounter.Linq.Visitors
 {
-    public class OrderByVisitor<TEntity> : StatelessVisitor<QueryBuilder<TEntity>>
+    internal class OrderByVisitor<TEntity> : StatelessVisitor<QueryBuilder<TEntity>>
     {
         public static OrderByVisitor<TEntity> Instance = new OrderByVisitor<TEntity>();
